chore: routing과 url constants 분리
